LayoutTests/imported/w3c/web-platform-tests/css/css-pseudo/text-selection-expected.txt

helloworld
helloworld
helloworld

FAIL Selection ending in ::before assert_equals: toString expected "hello" but got "o"
FAIL Selection contained in ::before assert_equals: toString expected "" but got "llo"
PASS Selection ending in ::marker
PASS Selection contained in ::marker
FAIL Selection ending in ::before-marker assert_equals: toString expected "hello" but got "o"
FAIL Selection contained in ::before-marker assert_equals: toString expected "" but got "llo"
